[LLAPI] add setLang and setScore (#2448)
* [LLAPI] add setLang and setScore

* pragma once for header
