+++ src/redisearch_api.c
@@ -36,7 +36,7 @@ IndexSpec* RediSearch_CreateIndex(const char* name, const RSIndexOptions* option
   if (options->score || options->lang) {
     spec->rule = rm_calloc(1, sizeof *spec->rule);
     spec->rule->score_default = options->score ? options->score : DEFAULT_SCORE;
-    spec->rule->lang_default = options->lang ? options->lang : DEFAULT_LANGUAGE;
+    spec->rule->lang_default = RSLanguage_Find(options->lang, 0);
   }
 
   spec->getValue = options->gvcb;
@@ -84,6 +84,13 @@ const char *RediSearch_IndexGetLanguage(IndexSpec* sp) {
     return RSLanguage_ToString(sp->rule->lang_default);
   }
   return RSLanguage_ToString(DEFAULT_LANGUAGE);
+}
+
+int RediSearch_ValidateLanguage(const char *lang) {
+  if (!lang || RSLanguage_Find(lang, 0) == RS_LANG_UNSUPPORTED) {
+    return REDISEARCH_ERR;
+  }
+  return REDISEARCH_OK;
 }
 
 RSFieldID RediSearch_CreateField(IndexSpec* sp, const char* name, unsigned types,
@@ -615,6 +622,22 @@ void RediSearch_IndexOptionsSetGCPolicy(RSIndexOptions* options, int policy) {
   options->gcPolicy = policy;
 }
 
+int RediSearch_IndexOptionsSetScore(RSIndexOptions* options, double score) {
+  if (score < 0 || score > 1) {
+    return REDISEARCH_ERR;
+  }
+  options->score = score;
+  return REDISEARCH_OK;
+}
+
+int RediSearch_IndexOptionsSetLanguage(RSIndexOptions* options, const char *lang) {
+  if (!lang || RediSearch_ValidateLanguage(lang) != REDISEARCH_OK) {
+    return REDISEARCH_ERR;
+  }
+  options->lang = lang;
+  return REDISEARCH_OK;
+}
+
 #define REGISTER_API(name)                                                          \
   if (RedisModule_ExportSharedAPI(ctx, "RediSearch_" #name, RediSearch_##name) !=   \
       REDISMODULE_OK) {                                                             \
@@ -686,10 +709,10 @@ int RediSearch_IndexInfo(RSIndex* sp, RSIdxInfo *info) {
   info->gcPolicy = sp->gc ? GC_POLICY_FORK : GC_POLICY_NONE;
   if (sp->rule) {
     info->score = sp->rule->score_default;
-    info->lang = sp->rule->lang_default;
+    info->lang = RSLanguage_ToString(sp->rule->lang_default);
   } else {
     info->score = DEFAULT_SCORE;
-    info->lang = DEFAULT_LANGUAGE;
+    info->lang = RSLanguage_ToString(DEFAULT_LANGUAGE);
   }
 
   info->numFields = sp->numFields;
