+++ src/stemmer.h
@@ -1,47 +1,13 @@
 #ifndef __RS_STEMMER_H__
 #define __RS_STEMMER_H__
-#include <stdlib.h>
+#include "language.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-typedef enum {
-  RS_LANG_ENGLISH = 0,
-  RS_LANG_ARABIC,
-  RS_LANG_BASQUE,
-  RS_LANG_CATALAN,
-  RS_LANG_CHINESE,
-  RS_LANG_DANISH,
-  RS_LANG_DUTCH,
-  RS_LANG_FINNISH,
-  RS_LANG_FRENCH,
-  RS_LANG_GERMAN,
-  RS_LANG_GREEK,
-  RS_LANG_HINDI,
-  RS_LANG_HUNGARIAN,
-  RS_LANG_ITALIAN,
-  RS_LANG_INDONESIAN,
-  RS_LANG_IRISH,
-  RS_LANG_LITHUANIAN,
-  RS_LANG_NEPALI,
-  RS_LANG_NORWEGIAN,
-  RS_LANG_PORTUGUESE,
-  RS_LANG_ROMANIAN,
-  RS_LANG_RUSSIAN,
-  RS_LANG_SPANISH,
-  RS_LANG_SWEDISH,
-  RS_LANG_TAMIL,
-  RS_LANG_TURKISH,
-  RS_LANG_ARMENIAN,
-  RS_LANG_SERBIAN,
-  RS_LANG_YIDDISH,
-  RS_LANG_UNSUPPORTED
-} RSLanguage;
-
 typedef enum { SnowballStemmer } StemmerType;
 
-#define DEFAULT_LANGUAGE RS_LANG_ENGLISH
 #define STEM_PREFIX '+'
 #define STEMMER_EXPANDER_NAME "stem"
 
@@ -63,10 +29,6 @@ typedef struct stemmer {
 Stemmer *NewStemmer(StemmerType type, RSLanguage language);
 
 int ResetStemmer(Stemmer *stemmer, StemmerType type, RSLanguage language);
-
-/* check if a language is supported by our stemmers */
-RSLanguage RSLanguage_Find(const char *language, size_t len);
-const char *RSLanguage_ToString(RSLanguage language);
 
 /* Get a stemmer expander instance for registering it */
 void RegisterStemmerExpander();
